+++ CHANGELOG.md
@@ -3,6 +3,8 @@ We use [semantic versioning][semver]
 # Next Release
 - [fix] Prevent "out of memory" in small JVMs: Don't cache test executions in memory
 - [breaking change] `--ignore-duplicates` (and `-d` option in convert tool) have been replaced with `--duplicates` option
+- [breaking change] `--filter` option in convert tool has been renamed to `--includes`
+- [breaking change] `--exclude` option in convert tool has been renamed to `--excludes`
 
 # 15.5.0
 - [feature] add TIA client library for integrating TIA in your custom test framework